chore: Moving go files to src folder
