--- day1/go.mod
@@ -1,3 +1,0 @@
-module github.com/truggeri/adventofcode2024/day1
-
-go 1.23.4
