--- day4/go.mod
@@ -1,3 +1,0 @@
-module github.com/truggeri/adventofcode2024/day4
-
-go 1.23.4
