+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/truggeri/adventofcode2024
+
+go 1.23.4
