+++ src/day2/day2_test.go
@@ -1,18 +1,18 @@
-package main
+package day2
 
 import (
 	"testing"
 )
 
 func TestSample(t *testing.T) {
-	input := `3   4
-4   3
-2   5
-1   3
-3   9
-3   3`
+	input := `7 6 4 2 1
+1 2 7 8 9
+9 7 6 2 1
+1 3 2 4 5
+8 6 4 4 1
+1 3 6 7 9`
 	result := Solve(input)
-	if result != 11 {
+	if result != 2 {
 		t.Errorf("Calculated solution was not expected")
 	}
 }
