+++ src/day3/day3.go
@@ -1,7 +1,6 @@
-package main
+package day3
 
 import (
-	"fmt"
 	"regexp"
 	"slices"
 	"strconv"
@@ -18,11 +17,6 @@ type mul struct {
 
 func (m mul) Eval() uint {
 	return m.x * m.y
-}
-
-func main() {
-	input := "xmul(2,4)%&mul[3,7]!@^do_not_mul(5,5)+mul(32,64]then(mul(11,8)mul(8,5))"
-	fmt.Println("Solution:", Solve(input))
 }
 
 func Solve(input string) uint {
