+++ src/day3/day3_test.go
@@ -1,18 +1,13 @@
-package main
+package day3
 
 import (
 	"testing"
 )
 
 func TestSample(t *testing.T) {
-	input := `7 6 4 2 1
-1 2 7 8 9
-9 7 6 2 1
-1 3 2 4 5
-8 6 4 4 1
-1 3 6 7 9`
+	input := "xmul(2,4)%&mul[3,7]!@^do_not_mul(5,5)+mul(32,64]then(mul(11,8)mul(8,5))"
 	result := Solve(input)
-	if result != 2 {
+	if result != 161 {
 		t.Errorf("Calculated solution was not expected")
 	}
 }
