+++ src/day4/day4_test.go
@@ -1,4 +1,4 @@
-package main
+package day4
 
 import (
 	"testing"
