+++ src/day5/day5.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"fmt"
 	"slices"
 	"strconv"
 	"strings"
@@ -10,38 +9,6 @@ import (
 type pageNumber uint
 type rule pageNumber
 type batch []pageNumber
-
-func main() {
-	input := `47|53
-97|13
-97|61
-97|47
-75|29
-61|13
-75|53
-29|13
-97|29
-53|29
-61|53
-97|53
-61|29
-47|13
-75|47
-97|75
-47|61
-75|61
-47|29
-75|13
-53|13
-
-75,47,61,53,29
-97,61,53,29,13
-75,29,13
-75,97,47,61,53
-61,13,29
-97,13,75,29,47`
-	fmt.Println("Solution:", Solve(input))
-}
 
 func Solve(input string) uint {
 	rules, batches := parseInput(input)
